requirements.txt
Flask==3.0.0
gunicorn==20.1.0
python-dotenv==1.0.0 